+++ paper/paper.bib
@@ -145,16 +145,6 @@
 }
 
 
-@inproceedings{stanvcin2019overview,
-  title={An overview and comparison of free {P}ython libraries for data mining and big data analysis},
-  author={Stan{\v{c}}in, Igor and Jovi{\'c}, Alan},
-  booktitle={2019 42nd International Convention on Information and Communication Technology, Electronics and Microelectronics (MIPRO)},
-  pages={977--982},
-  year={2019},
-  organization={IEEE}
-}
-
-
 @misc{chollet2015keras,
   title={Keras},
   author={Chollet, Fran\c{c}ois and others},
