### 2022-12-14 #5012 @ThomasMBury
Hi @danielskatz, thanks for helping with this. I've checked the 'missing DOIs' and they are correct for the papers I wish to cite. There is just one reference that I have removed from the bib file as I don't cite it in the paper.

Potential reviewers from the JOSS list could be
ry280888@gmail.com
kaur.saranjeet3@gmail.com
srobaina@ull.edu.es
canimus@gmail.com
