### 2022-12-14 #5012 @danielskatz
@ThomasMBury - there are still 5 possible DOIs that should be added to bibtex records - are you saying that none of these are correct?
