### 2022-12-14 #5012 @ThomasMBury
@danielskatz - the DOIs suggested by the bot are actually all correct. My .bib file does not contain any DOI entries - should I include the DOIs for each paper?
