### 2022-12-14 9c0e545
add dois to bib file
