+++ paper/paper.bib
@@ -1,163 +1,163 @@
+%% This BibTeX bibliography file was created using BibDesk.
+%% https://bibdesk.sourceforge.io/
+
+%% Created for Thomas Bury at 2022-12-14 12:22:07 -0500 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
 @article{scheffer2009early,
-  title={Early-warning signals for critical transitions},
-  author={Scheffer, Marten and Bascompte, Jordi and Brock, William A and Brovkin, Victor and Carpenter, Stephen R and Dakos, Vasilis and Held, Hermann and Van Nes, Egbert H and Rietkerk, Max and Sugihara, George},
-  journal={Nature},
-  volume={461},
-  number={7260},
-  pages={53--59},
-  year={2009},
-  publisher={Nature Publishing Group}
-}
+	author = {Scheffer, Marten and Bascompte, Jordi and Brock, William A and Brovkin, Victor and Carpenter, Stephen R and Dakos, Vasilis and Held, Hermann and Van Nes, Egbert H and Rietkerk, Max and Sugihara, George},
+	journal = {Nature},
+	number = {7260},
+	pages = {53--59},
+	publisher = {Nature Publishing Group},
+	title = {Early-warning signals for critical transitions},
+	volume = {461},
+	year = {2009},
+	doi = {10.1038/nature08227}}
 
 @article{clements2018indicators,
-  title={Indicators of transitions in biological systems},
-  author={Clements, Christopher F and Ozgul, Arpat},
-  journal={Ecology letters},
-  volume={21},
-  number={6},
-  pages={905--919},
-  year={2018},
-  publisher={Wiley Online Library}
-}
+	author = {Clements, Christopher F and Ozgul, Arpat},
+	journal = {Ecology letters},
+	number = {6},
+	pages = {905--919},
+	publisher = {Wiley Online Library},
+	title = {Indicators of transitions in biological systems},
+	volume = {21},
+	year = {2018},
+	doi = {10.1111/ele.12948}}
 
 @article{bury2020detecting,
-  title={Detecting and distinguishing tipping points using spectral early warning signals},
-  author={Bury, Thomas M and Bauch, Chris T and Anand, Madhur},
-  journal={Journal of the Royal Society Interface},
-  volume={17},
-  number={170},
-  pages={20200482},
-  year={2020},
-  publisher={The Royal Society}
-}
+	author = {Bury, Thomas M and Bauch, Chris T and Anand, Madhur},
+	journal = {Journal of the Royal Society Interface},
+	number = {170},
+	pages = {20200482},
+	publisher = {The Royal Society},
+	title = {Detecting and distinguishing tipping points using spectral early warning signals},
+	volume = {17},
+	year = {2020},
+	doi = {10.1098/rsif.2020.0482}}
 
 @article{bury2021deep,
-  title={Deep learning for early warning signals of tipping points},
-  author={Bury, Thomas M and Sujith, RI and Pavithran, Induja and Scheffer, Marten and Lenton, Timothy M and Anand, Madhur and Bauch, Chris T},
-  journal={Proceedings of the National Academy of Sciences},
-  volume={118},
-  number={39},
-  pages={e2106140118},
-  year={2021},
-  publisher={National Acad Sciences}
-}
+	author = {Bury, Thomas M and Sujith, RI and Pavithran, Induja and Scheffer, Marten and Lenton, Timothy M and Anand, Madhur and Bauch, Chris T},
+	journal = {Proceedings of the National Academy of Sciences},
+	number = {39},
+	pages = {e2106140118},
[…]
