+++ paper/paper.bib
@@ -85,12 +85,23 @@
 	year = {2020},
 	doi = {10.1038/s41586-020-2649-2}}
 
-@article{newville2016lmfit,
-	author = {Newville, Matthew and Stensitzki, Till and Allen, Daniel B and Rawlik, Michal and Ingargiola, Antonino and Nelson, Andrew},
-	journal = {Astrophysics Source Code Library},
-	pages = {ascl--1606},
-	title = {LMFIT: Non-linear least-square minimization and curve-fitting for {P}ython},
-	year = {2016}}
+
+
+@MISC{newville2016lmfit,
+       author = {{Newville}, Matthew and {Stensitzki}, Till and {Allen}, Daniel B. and {Rawlik}, Michal and {Ingargiola}, Antonino and {Nelson}, Andrew},
+        title = "{Lmfit: Non-Linear Least-Square Minimization and Curve-Fitting for Python}",
+     keywords = {Software},
+ howpublished = {Astrophysics Source Code Library, record ascl:1606.014},
+         year = 2016,
+        month = jun,
+          eid = {ascl:1606.014},
+        pages = {ascl:1606.014},
+archivePrefix = {ascl},
+       eprint = {1606.014},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2016ascl.soft06014N},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 
 @inproceedings{mckinney2010data,
 	author = {McKinney, Wes and others},
