+++ paper/paper.bib
@@ -86,7 +86,6 @@
 	doi = {10.1038/s41586-020-2649-2}}
 
 
-
 @MISC{newville2016lmfit,
        author = {{Newville}, Matthew and {Stensitzki}, Till and {Allen}, Daniel B. and {Rawlik}, Michal and {Ingargiola}, Antonino and {Nelson}, Andrew},
         title = "{Lmfit: Non-Linear Least-Square Minimization and Curve-Fitting for Python}",
@@ -101,7 +100,6 @@ archivePrefix = {ascl},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2016ascl.soft06014N},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
 
 @inproceedings{mckinney2010data,
 	author = {McKinney, Wes and others},
@@ -161,10 +159,12 @@ archivePrefix = {ascl},
 	doi = {10.48550/arXiv.1605.08695}}
 
 @misc{chollet2015keras,
-	author = {Chollet, Fran\c{c}ois and others},
-	howpublished = {\url{https://keras.io}},
-	title = {Keras},
-	year = {2015}}
+  title={Keras},
+  author={Chollet, Fran\c{c}ois and others},
+  year={2015},
+  publisher={GitHub},
+  howpublished={\url{https://github.com/fchollet/keras}},
+  }
 
 @online{pypl,
 	author = {PYPL},
