### 2022-12-14 bcca328
make keras a software entry
