+++ paper/paper.bib
@@ -158,7 +158,7 @@ archivePrefix = {ascl},
 	year = {2016},
 	doi = {10.48550/arXiv.1605.08695}}
 
-@misc{chollet2015keras,
+@software{chollet2015keras,
   title={Keras},
   author={Chollet, Fran\c{c}ois and others},
   year={2015},
