### 2022-12-14 #5012 @ThomasMBury
@danielskatz, I'm not able to find a DOI for the Keras software - I don't think it has one. Any suggestions?

The current bibtex entry for it is
@software{chollet2015keras,
  title={Keras},
  author={Chollet, Fran\c{c}ois and others},
  year={2015},
  publisher={GitHub},
  howpublished={\url{https://github.com/fchollet/keras}},
  }
