### 2022-12-14 #5012 @danielskatz
that's fine - this is an example where editorialbot is wrong, so you don't need to add a DOI.  JOSS does want DOIs where they exist, as they can be used as an index that leads to better citation tracking.
