### 2023-01-04 #5012 @mhu48
Hi,

I would be happy to help with the review of the submission! Thank you for
the opportunity. Please provide me with the link.

Sincerely,

Ming

On Wed, Jan 4, 2023 at 12:37 PM Øystein Sørensen ***@***.***>
wrote:

> 👋 @ranzhengcode <https://github.com/ranzhengcode> @vipinagrawal25
> <https://github.com/vipinagrawal25> @mikesha2
> <https://github.com/mikesha2> @mhu48 <https://github.com/mhu48>, would
> any of you be willing to review this submission for JOSS? We carry out our
> checklist-driven reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
