### 2023-01-05 #5038 @editorialbot
[REVIEW]: ewstools: A Python package for early warning signals of bifurcations in time series data
**Submitting author:** @ThomasMBury (<a href="http://orcid.org/0000-0003-1595-9444">Thomas M Bury</a>)
**Repository:** https://github.com/ThomasMBury/ewstools
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.1
