### 2023-02-10 #5038 @osorensen
@ThomasMBury, could you please update the pandas citation to those suggested here (preferably include both): https://pandas.pydata.org/about/citing.html
