+++ paper/paper.bib
@@ -101,7 +101,7 @@ archivePrefix = {ascl},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@InProceedings{ mckinney-proc-scipy-2010,
+@InProceedings{ mckinney2010data,
   author    = { {W}es {M}c{K}inney },
   title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
   booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
@@ -112,7 +112,7 @@ archivePrefix = {ascl},
 }
 
 @software{reback2020pandas,
-    author       = {The pandas development team},
+    author       = {{The pandas development team}},
     title        = {pandas-dev/pandas: Pandas},
     month        = feb,
     year         = 2020,
