### 2023-02-10 #5038 @ThomasMBury
Hi @osorensen, thank you for serving as the editor for this paper. I've completed the tasks above.

The new version of the software is v2.1.1.

The DOI on Zenodo is 10.5281/zenodo.7630022.

Thanks.
