### 2023-02-10 #5038 @ThomasMBury
I'm afraid I couldn't find a DOI for the Keras API - I hope this is ok. The bot didn't seem to mind the first time I submitted.
