+++ paper/paper.bib
@@ -53,7 +53,7 @@
 	doi = {10.1073/pnas.2106140118}}
 
 @article{dakos2012methods,
-	author = {Dakos, Vasilis and Carpenter, Stephen R and Brock, William A and Ellison, Aaron M and Guttal, Vishwesha and Ives, Anthony R and K{\'e}fi, Sonia and Livina, Valerie and Seekell, David A and van Nes, Egbert H and others},
+	author = {Dakos, Vasilis and Carpenter, Stephen R and Brock, William A and Ellison, Aaron M and Guttal, Vishwesha and Ives, Anthony R and K{\'e}fi, Sonia and Livina, Valerie and Seekell, David A and {van Nes}, Egbert H and others},
 	journal = {PloS one},
 	number = {7},
 	pages = {e41010},
@@ -88,7 +88,7 @@
 
 @MISC{newville2016lmfit,
        author = {{Newville}, Matthew and {Stensitzki}, Till and {Allen}, Daniel B. and {Rawlik}, Michal and {Ingargiola}, Antonino and {Nelson}, Andrew},
-        title = "{Lmfit: Non-Linear Least-Square Minimization and Curve-Fitting for Python}",
+        title = "Lmfit: Non-Linear Least-Square Minimization and Curve-Fitting for {P}ython",
      keywords = {Software},
  howpublished = {Astrophysics Source Code Library, record ascl:1606.014},
          year = 2016,
@@ -164,7 +164,7 @@ archivePrefix = {ascl},
 	author = {Abadi, Mart{\'\i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and others},
 	booktitle = {12th USENIX symposium on operating systems design and implementation (OSDI 16)},
 	pages = {265--283},
-	title = {Tensorflow: A system for large-scale machine learning},
+	title = {Tensor{F}low: A system for large-scale machine learning},
 	year = {2016},
 	doi = {10.48550/arXiv.1605.08695}}
 
