### 2023-02-11 #5038 @danielskatz
👋 @ThomasMBury - I'm working on the next steps for this submission, and after proofreading it, I have some suggested changes, in https://github.com/ThomasMBury/ewstools/pull/433  Please merge this, or let me know what you disagree with, then we can continue the process.
